1. Refresh of Realtime screen for 1 sec. 2. Added new fields in advance configuration screen. 3. Added background loading of faults. 4. Changes in the data formats of several fields in configuration screen. 5. Change in connection and disconnection behavior of the app, while coming back to advertising screen or going to background or powered off.

+++ LittleFuse/LittleFuse/Utility/LFUtilities.h
@@ -23,6 +23,6 @@
 
 + (NSString *)convertToHFormate:(float)val;
 
-+ (NSString *)convertToLFormate:(float)val;
++ (NSString *)convertToLFormate:(double)val;
 
 @end
